renaming variables and files
